backend/requirements.txt
numpy==1.26.4
openai==1.12.0
protobuf==4.25.2
pydantic==2.6.1
pypdf==4.0.1
python-dotenv==1.0.1
python-multipart==0.0.7
PyYAML==6.0.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
starlette==0.36.3
sympy==1.12
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.1
tqdm==4.66.1
typing-inspect==0.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
chromadb==0.4.15
InstructorEmbedding==1.0.1
autollm==0.1.10
docx2txt==0.8
pandas==2.2.0
langchain==0.1.13
langchain-community==0.0.29
pdfminer.six==20231228